+++ changelog.md
@@ -1,4 +1,5 @@
 * Background music
+* Sound effects
 * Improved UI
 * Improved tutorial
 * Improved color difference calculation (CIEDE2000)